moved bootstrappertests artifacts to fakes folder
