+++ tests/BoltOn.Tests/Bootstrapping/Fakes/BootstrapperRegistrationTasksHelper.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace BoltOn.Tests.Bootstrapping.Fakes
+{
+    public class BootstrapperRegistrationTasksHelper
+    {
+        public static List<string> Tasks { get; set; } = new List<string>();
+    }
+}
