+++ tests/BoltOn.Tests/Bootstrapping/Fakes/ClassWithInjectedDependency.cs
@@ -1,0 +1,16 @@
+﻿namespace BoltOn.Tests.Bootstrapping.Fakes
+{
+    public class ClassWithInjectedDependency
+    {
+        public ClassWithInjectedDependency(ITestService testService)
+        {
+            Name = testService.GetName();
+        }
+
+        public string Name
+        {
+            get;
+            set;
+        }
+    }
+}
