+++ tests/BoltOn.Tests/Bootstrapping/Fakes/TestBootstrapperPostRegistrationTask.cs
@@ -1,0 +1,12 @@
+﻿using BoltOn.Bootstrapping;
+
+namespace BoltOn.Tests.Bootstrapping.Fakes
+{
+    public class TestBootstrapperPostRegistrationTask : IPostRegistrationTask
+    {
+        public void Run()
+        {
+            BootstrapperRegistrationTasksHelper.Tasks.Add($"Executed {GetType().Name}");
+        }
+    }
+}
