+++ tests/BoltOn.Tests/Bootstrapping/Fakes/TestService.cs
@@ -1,0 +1,15 @@
+﻿namespace BoltOn.Tests.Bootstrapping.Fakes
+{
+	public interface ITestService
+	{
+		string GetName();
+	}
+
+	public class TestService : ITestService
+    {
+        public string GetName()
+        {
+            return "test";
+        }
+    }
+}
